% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder_process.R
\name{gapmider_rel_pop}
\alias{gapmider_rel_pop}
\title{Calculate population relative to a baseline year}
\usage{
gapmider_rel_pop(gapminder_df, base_year = 1952)
}
\arguments{
\item{gapminder_df}{dataframe of gapminder data}

\item{base_year}{numeric. Year to be taken as baseline for the calculation of 
of relative population change}
}
\value{
gapminder_df with relative population added as column pop_rel
}
\description{
Calculate population at different timepoints, for each country in the gapminder 
dataset relative to a baseline year. Append result to the dataframe as a new column.
}
